PM-367 - Pre PR Modifications - Moved generation to Abstract BGPSession Model - Added property value into BGPSession to check the session_type - Removed Py3 annotations
