+++ peering/models/abstracts.py
@@ -1,5 +1,6 @@
 import ipaddress
 import logging
+import uuid
 
 from django.db import models
 from django.utils.safestring import mark_safe
@@ -87,6 +88,12 @@ class BGPSession(ChangeLoggedModel, TaggableModel, PolicyMixin):
     )
     bgp_state = models.CharField(
         max_length=50, choices=BGPState.choices, blank=True, null=True
+    )
+    service_reference = models.CharField(
+        max_length=255,
+        unique=True,
+        blank=True,
+        help_text="Optional: Internal Service Reference (will auto generate if left blank)",
     )
     received_prefix_count = models.PositiveIntegerField(blank=True, default=0)
     advertised_prefix_count = models.PositiveIntegerField(blank=True, default=0)
@@ -177,6 +184,47 @@ class BGPSession(ChangeLoggedModel, TaggableModel, PolicyMixin):
         if commit:
             self.save()
         return True
+
+    @property
+    def session_type(self):
+        """
+        Determine what type of BGPSession this class has
+        been inherited into
+
+        Return:
+            str: ixp or direct
+        """
+        return "ixp" if hasattr(self, "ixp_connection") else "direct"
+
+    def generate_service_ref(self):
+        """
+        Generate a Unique Service Reference for an IX/Direct
+        session from Local ASN with 6 digit hex UUID.
+
+        Example: IX9268-FD130FS/IX<asn>-<hex>S
+        Example: D9268-4CD335S/D<asn>-<hex>S
+
+        Return:
+            str: Service Reference
+        """
+        if self.session_type == "ixp":
+            asn = (
+                self.ixp_connection.internet_exchange_point.local_autonomous_system.asn
+            )
+            return "IX{0}-{1}S".format(asn, uuid.uuid4().hex[:6].upper())
+        else:
+            asn = self.local_autonomous_system.asn
+            return "D{0}-{1}S".format(asn, uuid.uuid4().hex[:6].upper())
+
+    def save(self, *args, **kwargs):
+        """
+        Overwrite Model Save to generate a Unique Service Reference
+        if left blank on Save/Update
+        """
+        if not self.service_reference:
+            self.service_reference = self.generate_service_ref()
+
+        return super().save(*args, **kwargs)
 
 
 class Template(ChangeLoggedModel, TaggableModel):
