+++ templates/peering/directpeeringsession/details.html
@@ -7,6 +7,10 @@
       <div class="card-header"><strong>Peering Session Details</strong></div>
       <table class="card-body table table-hover attr-table mb-0">
         <tr>
+          <td>Service ID</td>
+          <td>{{ instance.service_reference|as_link }}</td>
+        </tr>
+        <tr>
           <td>Relationship</td>
           <td>{{ instance.get_relationship_display }}</td>
         </tr>
@@ -64,10 +68,6 @@
           <td>{{ instance.router|as_link }}</td>
         </tr>
         {% endif %}
-        <tr>
-          <td>Service ID</td>
-          <td>{{ instance.service_reference|as_link }}</td>
-        </tr>
       </table>
     </div>
     <div class="card">
